Add files via upload
Since the original dataset was very big in size. I was unable to upload it. However, the "new_train.csv" file is the outcome of all the things that I've tried in "TCC1.ipynb" to make the dataset usable on low end machines and also deal with the class imbalance problem as well.